bigquery-loader-service/requirements.txt
google-cloud-bigquery==3.14.0
google-cloud-storage==2.14.0
pyarrow==14.0.1
pandas==2.1.4